Move guice dependent classes to com.netflix.governator.guice.
