+++ governator-core/src/main/java/com/netflix/governator/LifecycleListener.java
@@ -1,6 +1,5 @@
 package com.netflix.governator;
 
-
 /**
  * Contract for a callback to be invoked once the Injector is shut down.
  * LifecycleEvent should be reserved for framework use only.  All application
